+++ ivy/functional/frontends/paddle/tensor/math.py
@@ -210,24 +210,6 @@ def exp(x, name=None):
     return ivy.exp(x)
 
 
-@with_supported_dtypes(
-    {
-        "2.4.2 and below": (
-            "float32",
-            "float64",
-            "int32",
-            "int64",
-            "complex64",
-            "complex128",
-        )
-    },
-    "paddle",
-)
-@to_ivy_arrays_and_back
-def cumprod(x, dim=None, dtype=None, name=None):
-    return ivy.cumprod(x, axis=dim, dtype=dtype)
- 
-
 @with_unsupported_dtypes({"2.5.0 and below": ("float16", "bfloat16")}, "paddle")
 @to_ivy_arrays_and_back
 def reciprocal(x, name=None):
