+++ ivy/functional/frontends/torch/random_sampling.py
@@ -66,7 +66,7 @@ def randint(
     dtype=None,
     layout=None,
     device=None,
-    requires_grad=False,
+    requires_grad=False
 ):
     seed = generator.initial_seed() if generator is not None else None
     return ivy.randint(
@@ -82,25 +82,16 @@ def randint(
 
 @to_ivy_arrays_and_back
 def rand(
-    *size,
+    size,
+    *,
     generator=None,
     out=None,
     dtype=None,
     layout=None,
     device=None,
     requires_grad=False,
-    pin_memory=False,
-    **kwargs,
+    pin_memory=False
 ):
-    if not size and "size" not in kwargs:
-        raise ValueError("Missing 1 required positional/keyword argument: size")
-    size = kwargs["size"] if not size else size
-    if (
-        isinstance(size, (list, tuple))
-        and len(size) == 1
-        and isinstance(size[0], (list, tuple))
-    ):
-        size = size[0]
     seed = generator.initial_seed() if generator is not None else None
     return ivy.random_uniform(
         shape=size,
@@ -134,7 +125,7 @@ def rand_like(
     layout=None,
     device=None,
     requires_grad=False,
-    memory_format=False,
+    memory_format=False
 ):
     shape = input.shape
     if not dtype:
@@ -149,25 +140,16 @@ def rand_like(
 
 @to_ivy_arrays_and_back
 def randn(
-    *size,
+    size,
+    *,
     generator=None,
     out=None,
     dtype=None,
     layout=None,
     device=None,
     requires_grad=False,
-    pin_memory=False,
-    **kwargs,
+    pin_memory=False
 ):
-    if not size and "size" not in kwargs:
-        raise ValueError("Missing 1 required positional/keyword argument: size")
-    size = kwargs["size"] if not size else size
-    if (
-        isinstance(size, (list, tuple))
-        and len(size) == 1
-        and isinstance(size[0], (list, tuple))
-    ):
-        size = size[0]
     seed = generator.initial_seed() if generator is not None else None
     return ivy.random_normal(
         shape=size,
@@ -186,7 +168,7 @@ def randn_like(
     layout=None,
     device=None,
     requires_grad=False,
-    memory_format=None,
+    memory_format=None
 ):
     shape = input.shape
     if not dtype:
@@ -224,7 +206,7 @@ def randperm(
     layout=None,
     device=None,
     requires_grad=False,
-    pin_memory=False,
+    pin_memory=False
 ):
     seed = generator.initial_seed() if generator is not None else None
     arr = ivy.arange(n, device=device, dtype=dtype)
