+++ ivy_tests/test_ivy/test_frontends/test_paddle/test_tensor/test_paddle_manipulation.py
@@ -246,49 +246,3 @@ def test_paddle_tile(
         x=x[0],
         repeat_times=repeats,
     )
-
-
-# split
-@st.composite
-def _split_helper(draw):
-    dtypes, values, shape = draw(
-        helpers.dtype_and_values(
-            available_dtypes=helpers.get_dtypes("valid"),
-            min_num_dims=2,
-            max_num_dims=4,
-            min_dim_size=2,
-            max_dim_size=4,
-            ret_shape=True,
-        )
-    )
-    axis = draw(st.sampled_from(range(len(shape))))
-    num_eles = shape[axis]
-    splits = [i for i in range(1, num_eles + 1) if num_eles % i == 0]
-    num_splits = draw(st.sampled_from(splits))
-    return dtypes, values, num_splits, axis
-
-
-@handle_frontend_test(
-    fn_tree="paddle.split",
-    dt_x_num_splits_axis=_split_helper(),
-    test_with_out=st.just(False),
-)
-def test_paddle_split(
-    *,
-    dt_x_num_splits_axis,
-    on_device,
-    fn_tree,
-    frontend,
-    test_flags,
-):
-    input_dtypes, x, num_splits, axis = dt_x_num_splits_axis
-    helpers.test_frontend_function(
-        input_dtypes=input_dtypes,
-        frontend=frontend,
-        test_flags=test_flags,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-        num_or_sections=num_splits,
-        axis=axis,
-    )
