+++ ivy_tests/test_ivy/test_frontends/test_paddle/test_tensor/test_paddle_math.py
@@ -938,41 +938,6 @@ def test_paddle_reciprocal(
         x=x[0],
     )
 
-    
-# cumprod
-@handle_frontend_test(
-    fn_tree="paddle.tensor.math.cumprod",
-    dtype_and_x=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("valid"),
-        safety_factor_scale="log",
-        large_abs_safety_factor=6,
-        small_abs_safety_factor=5,
-        min_num_dims=1,
-        valid_axis=True,
-        force_int_axis=True,
-    ),
-)
-def test_paddle_cumprod(
-    *,
-    dtype_and_x,
-    on_device,
-    fn_tree,
-    frontend,
-    test_flags,
-):
-    input_dtype, x, axis = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        frontend=frontend,
-        test_flags=test_flags,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-        dim=axis,
-        rtol=1e-01,
-        atol=1e-01,
-    )
-
 
 # gcd
 @handle_frontend_test(
