+++ ivy_tests/test_ivy/test_frontends/test_torch/test_random_sampling.py
@@ -177,16 +177,13 @@ def test_torch_randint(
     fn_tree="torch.rand",
     dtype=helpers.get_dtypes("float", full=False),
     size=helpers.get_shape(
-        min_num_dims=1,
+        min_num_dims=0,
         max_num_dims=5,
-        min_dim_size=1,
+        min_dim_size=0,
         max_dim_size=10,
     ),
 )
 def test_torch_rand(*, dtype, size, frontend, fn_tree, test_flags):
-    size = {f"size{i}": size[i] for i in range(len(size))}
-    test_flags.num_positional_args = len(size)
-
     def call():
         return helpers.test_frontend_function(
             input_dtypes=dtype,
@@ -194,7 +191,7 @@ def test_torch_rand(*, dtype, size, frontend, fn_tree, test_flags):
             test_values=False,
             fn_tree=fn_tree,
             test_flags=test_flags,
-            **size,
+            size=size,
         )
 
     ret = call()
@@ -308,16 +305,13 @@ def test_torch_rand_like(dtype_and_x, dtype, *, frontend, fn_tree, test_flags):
     fn_tree="torch.randn",
     dtype=helpers.get_dtypes("float", full=False),
     size=helpers.get_shape(
-        min_num_dims=1,
+        min_num_dims=0,
         max_num_dims=5,
-        min_dim_size=1,
+        min_dim_size=0,
         max_dim_size=10,
     ),
 )
 def test_torch_randn(*, dtype, size, frontend, fn_tree, test_flags):
-    size = {f"size{i}": size[i] for i in range(len(size))}
-    test_flags.num_positional_args = len(size)
-
     def call():
         return helpers.test_frontend_function(
             input_dtypes=dtype,
@@ -325,7 +319,7 @@ def test_torch_randn(*, dtype, size, frontend, fn_tree, test_flags):
             test_values=False,
             fn_tree=fn_tree,
             test_flags=test_flags,
-            **size,
+            size=size,
         )
 
     ret = call()
