--- ivy_tests/test_ivy/test_functional/test_experimental/test_nn/test_losses.py
@@ -1,60 +1,0 @@
-# global
-from hypothesis import strategies as st
-
-# local
-import ivy_tests.test_ivy.helpers as helpers
-from ivy_tests.test_ivy.helpers import handle_test
-
-
-# log_poisson_loss
-@handle_test(
-    fn_tree="functional.ivy.experimental.log_poisson_loss",
-    dtype_and_targets=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_value=0,
-        max_value=3,
-        allow_inf=False,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=3,
-    ),
-    dtype_and_log_input=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        small_abs_safety_factor=4,
-        safety_factor_scale="log",
-        min_value=0,
-        max_value=3,
-        allow_inf=False,
-        exclude_min=True,
-        exclude_max=True,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=3,
-    ),
-    compute_full_loss=st.sampled_from([True, False]),
-    test_with_out=st.just(False),
-)
-def test_log_poisson_loss(
-    dtype_and_targets,
-    dtype_and_log_input,
-    compute_full_loss,
-    test_flags,
-    backend_fw,
-    fn_name,
-    on_device,
-    ground_truth_backend,
-):
-    targets_dtype, targets = dtype_and_targets
-    log_input_dtype, log_input = dtype_and_log_input
-    helpers.test_function(
-        ground_truth_backend=ground_truth_backend,
-        input_dtypes=targets_dtype + log_input_dtype,
-        test_flags=test_flags,
-        fw=backend_fw,
-        fn_name=fn_name,
-        on_device=on_device,
-        targets=targets[0],
-        log_input=log_input[0],
-        compute_full_loss=compute_full_loss,
-        atol_=1e-2,
-    )
